+++ src/Oro/Bundle/FrontendBundle/Resources/public/blank/js/app/views/frontend-search-boolean-filter.js
@@ -1,12 +1,12 @@
 define(function(require) {
     'use strict';
 
-    var FrontendBooleanFilter;
+    var FrontendSearchBooleanFilter;
     var _ = require('underscore');
     var MultiSelectFilter = require('oro/filter/multiselect-filter');
     var FilterCountHelper = require('orofrontend/js/app/filter-count-helper');
 
-    FrontendBooleanFilter = MultiSelectFilter.extend(_.extend({}, FilterCountHelper, {
+    FrontendSearchBooleanFilter = MultiSelectFilter.extend(_.extend({}, FilterCountHelper, {
         /**
          * @inheritDoc
          */
@@ -24,18 +24,18 @@ define(function(require) {
          * @inheritDoc
          */
         constructor: function FrontendBooleanFilter() {
-            FrontendBooleanFilter.__super__.constructor.apply(this, arguments);
+            FrontendSearchBooleanFilter.__super__.constructor.apply(this, arguments);
         },
 
         /**
          * @inheritDoc
          */
         getTemplateData: function() {
-            var templateData = FrontendBooleanFilter.__super__.getTemplateData.apply(this, arguments);
+            var templateData = FrontendSearchBooleanFilter.__super__.getTemplateData.apply(this, arguments);
 
             return this.filterTemplateData(templateData);
         }
     }));
 
-    return FrontendBooleanFilter;
+    return FrontendSearchBooleanFilter;
 });