+++ src/Oro/Bundle/FrontendBundle/Resources/public/blank/js/app/views/frontend-select-filter.js
@@ -48,11 +48,6 @@ define(function(require) {
         containerSelector: '.filter-criteria-selector',
 
         /**
-         * @inheritDoc
-         */
-        populateDefault: false,
-
-        /**
          * @property {Object}
          */
         listen: {